+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-FJJb1985.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-CHe00_62.js";import{_ as a}from"./mermaid.core-Dj6oz06I.js";import"./chunk-5VM5RSS4-CIhCp1_X.js";import"./chunk-XXDRQBXY-BybM-WU8.js";import"./chunk-POPQ4Y6H-DKvUDWOQ.js";import"./chunk-F27PBJKO-WKxf_azs.js";import"./channel-CmZ_u0_7.js";import"./index-B2L_mFhI.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};